Time: 130 ms (76.89%) | Memory: 111.1 MB (92.67%) - LeetSync

+++ 295-find-median-from-data-stream/find-median-from-data-stream.java
@@ -1,16 +1,16 @@
 class MedianFinder {
-    PriorityQueue<Double> left;
-    PriorityQueue<Double> right;
+    PriorityQueue<Integer> left;
+    PriorityQueue<Integer> right;
     public MedianFinder() {
         left = new PriorityQueue<>(Collections.reverseOrder());
         right = new PriorityQueue<>();
     }
     
     public void addNum(int num) {
-        if(left.size() == 0) left.add((double)num);
+        if(left.size() == 0) left.add(num);
         else{
-            if(num<left.peek()) left.add((double)num);
-            else right.add((double)num);
+            if(num<left.peek()) left.add(num);
+            else right.add(num);
         }
 
         if(left.size() == right.size()+2){
